refactor(frontdesk): split server.go and store.go into cohesive files
Both files had grown large (server.go 1259 lines, store.go 1015). Split
each along its existing banner sections into same-package siblings. Pure
code relocation: no behavior, signature, or logic changes.

server.go (1259 -> 298): extracted server_members.go, server_settings.go
(settings + auto-sync + rearm), server_status.go (events/SSE/version/
observability/TOTP cache/emit), server_alerts.go, and httputil.go
(writeJSON/writeError/decodeJSON/spaHandler and small helpers). Core keeps
ServerConfig, Server, NewServer, buildRouter, auth middleware, traefik handler.

store.go (1015 -> 246): extracted store_members.go, store_settings.go
(settings + auto-sync), store_fleet.go, store_events.go, and store_helpers.go
(scanners/normalizeMemberURL/sentinels). Core keeps the types plus
Open/migrate/applyMigration.

Verified: go build/vet clean, golangci-lint 0 issues, all 191 frontdesk
tests pass.
