+++ internal/frontdesk/server.go
@@ -1,17 +1,11 @@
 package frontdesk
 
 import (
-	"context"
 	"crypto/subtle"
-	"encoding/json"
-	"errors"
-	"fmt"
 	"io/fs"
 	"net/http"
-	"strconv"
 	"strings"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	"github.com/go-chi/chi/v5"
@@ -21,10 +15,8 @@ import (
 	"github.com/hugalafutro/model-hotel/internal/admin"
 	"github.com/hugalafutro/model-hotel/internal/adminauth"
 	"github.com/hugalafutro/model-hotel/internal/alert"
-	"github.com/hugalafutro/model-hotel/internal/auth"
 	"github.com/hugalafutro/model-hotel/internal/debuglog"
 	"github.com/hugalafutro/model-hotel/internal/events"
-	"github.com/hugalafutro/model-hotel/internal/otelexport"
 	"github.com/hugalafutro/model-hotel/internal/totp"
 	"github.com/hugalafutro/model-hotel/internal/util"
 	"github.com/hugalafutro/model-hotel/internal/webauthn"
@@ -303,958 +295,4 @@ func (s *Server) handleTraefikConfig(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	writeJSON(w, http.StatusOK, BuildTraefikConfig(members, set))
-}
-
-// ---------------------------------------------------------------------------
-// Members
-// ---------------------------------------------------------------------------
-
-// memberView is a member plus its live poller status for the Members tab.
-type memberView struct {
-	*Member
-	Status MemberStatus `json:"status"`
-}
-
-func (s *Server) listMembers(w http.ResponseWriter, r *http.Request) {
-	members, err := s.store.ListMembers(r.Context())
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	snap := s.poller.Snapshot()
-	views := make([]memberView, len(members))
-	for i, m := range members {
-		views[i] = memberView{Member: m, Status: snap[m.ID]}
-	}
-	writeJSON(w, http.StatusOK, views)
-}
-
-type createMemberRequest struct {
-	Name  string `json:"name"`
-	URL   string `json:"url"`
-	Token string `json:"token"`
-}
-
-// memberResponse is a Member plus an optional, non-fatal warning surfaced after
-// an add/edit when the admin token could not be confirmed (the member was
-// offline, or answered without a 200). The frontend toasts token_warning when
-// it is present; a token the member positively refused is a 400 instead.
-type memberResponse struct {
-	*Member
-	TokenWarning string `json:"token_warning,omitempty"`
-}
-
-func (s *Server) createMember(w http.ResponseWriter, r *http.Request) {
-	var req createMemberRequest
-	if !decodeJSON(w, r, &req) {
-		return
-	}
-	// A member is only added once it replies and verifies: the admin token is
-	// required (there is no way to confirm the host's identity or fleet role
-	// without it), the host must answer the authenticated probe, and it must not
-	// already be the fleet primary reached under a different URL. This keeps the
-	// member list to genuine, distinct model-hotel instances and stops the primary
-	// (the config source of truth) from being re-added as an ordinary member.
-	if strings.TrimSpace(req.Token) == "" {
-		writeCodedError(w, http.StatusBadRequest, "token_required",
-			"an admin token is required to add a member: Front Desk uses it to verify the host and confirm its fleet role before adding it")
-		return
-	}
-	m, err := s.store.CreateMember(r.Context(), req.Name, req.URL, req.Token)
-	if err != nil {
-		// Map the two validation failures the add form routes on to stable codes;
-		// everything else falls back to the shared plain-text writeError.
-		switch {
-		case errors.Is(err, ErrDuplicateURL):
-			writeCodedError(w, http.StatusBadRequest, "duplicate", err.Error())
-		case errors.Is(err, ErrInsecureURL):
-			writeCodedError(w, http.StatusBadRequest, "insecure_url", err.Error())
-		default:
-			writeError(w, err)
-		}
-		return
-	}
-	// rollback removes the just-created row when a verification step fails, so a
-	// rejected add leaves no half-added member (and no duplicate-URL wall on retry).
-	rollback := func(code, userMsg string, status int) {
-		if delErr := s.store.DeleteMember(r.Context(), m.ID); delErr != nil {
-			writeCodedError(w, http.StatusInternalServerError, "rollback_failed",
-				fmt.Sprintf("%s Rolling back the add also failed (%v); remove it from the Members list and try again.", userMsg, delErr))
-			return
-		}
-		writeCodedError(w, status, code, userMsg)
-	}
-
-	// Verify the token against the (now canonical) member URL. Unlike an edit, an
-	// add requires a positive reply: an unreachable host or a refused/unexpected
-	// response blocks the add rather than warning, so only live, verified members
-	// enter the list.
-	p := s.probeMemberToken(r.Context(), m.URL, req.Token)
-	if !p.valid {
-		switch {
-		case !p.reached:
-			rollback("unreachable", "Front Desk could not reach this member to verify it. Check the URL and that the host is running, then try again.", http.StatusBadRequest)
-		case p.rejected():
-			rollback("token_rejected", fmt.Sprintf("This member rejected the admin token (HTTP %d). Double-check the token and try again.", p.status), http.StatusBadRequest)
-		default:
-			rollback("unverified", fmt.Sprintf("This host did not verify as a Front Desk member (HTTP %d). Check the URL points at a model-hotel instance and the token is correct.", p.status), http.StatusBadRequest)
-		}
-		return
-	}
-	// Read the host's identity once: whether it self-reports as the fleet primary,
-	// and its stable instance_id. Both are URL-independent, so they catch the same
-	// physical instance reached under a different address. The token probe above
-	// already confirmed the host is live and genuine, so a failure to read its
-	// identity here is anomalous: rather than fail open (which could admit the
-	// primary or a duplicate under a new URL), block the add and let the operator
-	// retry once the host answers /api/system cleanly.
-	isPrimary, instanceID, identOK := s.memberIdentity(r.Context(), m.URL, req.Token)
-	if !identOK {
-		rollback("identity_unverified", "Front Desk verified the admin token but could not read this host's fleet identity (/api/system) to confirm it is not the fleet primary or an existing member. Check the host and try again.", http.StatusBadRequest)
-		return
-	}
-	// Reject the fleet primary re-added under a different URL. Only one primary
-	// exists, so a host self-reporting is_primary is that primary reached under
-	// another address.
-	if isPrimary {
-		rollback("already_primary", "This host is already the fleet primary (the config source of truth), reached under a different address. It cannot also be added as a member.", http.StatusConflict)
-		return
-	}
-	// Reject a host that is already a member under a different URL: compare its
-	// instance_id against every other member. Any member whose id we do not yet
-	// know is probed once and backfilled, so this stays correct even for members
-	// added before instance identity existed. A pre-056 host that exposes no
-	// instance_id skips dedup (there is nothing to compare); it is the one
-	// residual gap, and adds now require a token anyway.
-	if instanceID != "" {
-		dup, derr := s.instanceAlreadyMember(r.Context(), m.ID, instanceID)
-		if derr != nil {
-			rollback("verify_failed", "Front Desk could not verify whether this host is already a member. Try again.", http.StatusInternalServerError)
-			return
-		}
-		if dup {
-			rollback("already_member", "This host is already a member (added under a different address). Remove the existing entry first if you want to re-add it.", http.StatusConflict)
-			return
-		}
-		// Persist the learned identity so future adds can dedup against this
-		// member without re-probing it. A failure to record it would leave the
-		// member half-registered (present but un-deduplicable), so roll the add
-		// back rather than let a duplicate slip in under a different URL later.
-		if err := s.store.SetMemberInstanceID(r.Context(), m.ID, instanceID); err != nil {
-			debuglog.Warn("frontdesk: could not store member instance id", "member", m.ID, "error", err)
-			rollback("verify_failed", "Front Desk verified this host but could not record its identity. Try again.", http.StatusInternalServerError)
-			return
-		}
-	}
-
-	// A newly added member with a valid token is stale relative to the primary;
-	// re-arm auto-sync so the next tick brings it in line (no-op when disabled).
-	s.rearmAutoSync(r.Context())
-	s.emit(r.Context(), Event{
-		Type: "member.added", Severity: "info", Source: "frontdesk",
-		Message: m.Name + " added", MemberID: m.ID,
-		Metadata: map[string]any{"url": m.URL},
-	})
-	writeJSON(w, http.StatusCreated, memberResponse{Member: m})
-}
-
-type patchMemberRequest struct {
-	Name  *string `json:"name,omitempty"`
-	Token *string `json:"token,omitempty"` // "" clears the stored token
-}
-
-func (s *Server) patchMember(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-	var req patchMemberRequest
-	if !decodeJSON(w, r, &req) {
-		return
-	}
-	if req.Name != nil {
-		if err := s.store.RenameMember(r.Context(), id, *req.Name); err != nil {
-			writeError(w, err)
-			return
-		}
-	}
-	var tokenWarning string
-	if req.Token != nil {
-		// Verify a non-empty new token before storing it, so a refused token is
-		// rejected now rather than persisted. Clearing the token ("") never probes.
-		if *req.Token != "" {
-			m0, err := s.store.GetMember(r.Context(), id)
-			if err != nil {
-				writeError(w, err)
-				return
-			}
-			p := s.probeMemberToken(r.Context(), m0.URL, *req.Token)
-			if p.rejected() {
-				http.Error(w, fmt.Sprintf("This member rejected the admin token (HTTP %d). Double-check the token and try again.", p.status), http.StatusBadRequest)
-				return
-			}
-			tokenWarning = p.warning()
-		}
-		if err := s.store.SetMemberToken(r.Context(), id, *req.Token); err != nil {
-			writeError(w, err)
-			return
-		}
-		if *req.Token != "" {
-			// The member just gained an admin token: it is now syncable but stale, so
-			// re-arm auto-sync to converge it (no-op when disabled).
-			s.rearmAutoSync(r.Context())
-		}
-	}
-	m, err := s.store.GetMember(r.Context(), id)
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	writeJSON(w, http.StatusOK, memberResponse{Member: m, TokenWarning: tokenWarning})
-}
-
-func (s *Server) deleteMember(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-	m, err := s.store.GetMember(r.Context(), id)
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	// The fleet primary is the config source of truth and cannot be removed here
-	// at all: changing it goes through the Fleet Sync wizard (a token-gated
-	// repoint). The primary-status check and the delete run as one atomic SQL
-	// statement inside DeleteMemberIfNotPrimary, so a concurrent repoint cannot
-	// race past the check.
-	applied, err := s.store.DeleteMemberIfNotPrimary(r.Context(), id)
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	if !applied {
-		http.Error(w, "this host is the fleet primary (the config source of truth); change the primary from the Fleet Sync wizard before removing it", http.StatusConflict)
-		return
-	}
-	s.emit(r.Context(), Event{
-		Type: "member.removed", Severity: "info", Source: "frontdesk",
-		Message: m.Name + " removed", MemberID: m.ID,
-	})
-	w.WriteHeader(http.StatusNoContent)
-}
-
-type memberStateRequest struct {
-	State MemberState `json:"state"`
-}
-
-func (s *Server) setMemberState(w http.ResponseWriter, r *http.Request) {
-	id := chi.URLParam(r, "id")
-	var req memberStateRequest
-	if !decodeJSON(w, r, &req) {
-		return
-	}
-	if err := s.store.SetMemberState(r.Context(), id, req.State); err != nil {
-		writeError(w, err)
-		return
-	}
-	m, err := s.store.GetMember(r.Context(), id)
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	severity := "info"
-	if req.State == StateDrained {
-		severity = "warning"
-	}
-	s.emit(r.Context(), Event{
-		Type: "member.state_changed", Severity: severity, Source: "frontdesk",
-		Message: m.Name + " set to " + string(req.State), MemberID: m.ID,
-		Metadata: map[string]any{"state": string(req.State)},
-	})
-	writeJSON(w, http.StatusOK, m)
-}
-
-// ---------------------------------------------------------------------------
-// Settings
-// ---------------------------------------------------------------------------
-
-func (s *Server) getSettings(w http.ResponseWriter, r *http.Request) {
-	set, err := s.store.GetSettings(r.Context())
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	// Never expose the encrypted Apprise target or OIDC client secret: replace a
-	// stored secret with a mask the UI can echo back unchanged to preserve it.
-	if set.AlertAppriseTargets != "" {
-		set.AlertAppriseTargets = alertMaskValue
-	}
-	if set.OidcClientSecret != "" {
-		set.OidcClientSecret = alertMaskValue
-	}
-	writeJSON(w, http.StatusOK, set)
-}
-
-func (s *Server) putSettings(w http.ResponseWriter, r *http.Request) {
-	// Partial-merge onto the current row: decode the request body on top of the
-	// stored settings so a field the caller omits is preserved, not zeroed. The
-	// polling form and the Alerts panel each PUT only the fields they own, so
-	// neither clobbers the other; and an older client that never sends the alert
-	// fields can no longer wipe the encrypted target. The target is presented as
-	// the mask first, so an omitted or echoed value both resolve to "preserve"
-	// (and the raw ciphertext is never re-encrypted into itself).
-	//
-	// The read-merge-write is serialized: two concurrent PUTs (e.g. both panels at
-	// once) must not both read the same snapshot and have the later write restore
-	// the other's pre-merge values. putSettings is the only writer of this row.
-	s.settingsMu.Lock()
-	defer s.settingsMu.Unlock()
-
-	set, err := s.store.GetSettings(r.Context())
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	if set.AlertAppriseTargets != "" {
-		set.AlertAppriseTargets = alertMaskValue
-	}
-	if set.OidcClientSecret != "" {
-		set.OidcClientSecret = alertMaskValue
-	}
-	if !decodeJSON(w, r, &set) {
-		return
-	}
-	// Resolve the Apprise target secret before storing: a masked submission keeps
-	// the existing ciphertext, a new value is encrypted at rest, a blank clears it.
-	resolved, err := s.resolveAlertTarget(r.Context(), set.AlertAppriseTargets)
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	set.AlertAppriseTargets = resolved
-	// The OIDC client secret follows the same mask/encrypt/preserve contract.
-	oidcSecret, err := s.resolveSecret(r.Context(), set.OidcClientSecret, func(cur Settings) string {
-		return cur.OidcClientSecret
-	})
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	set.OidcClientSecret = oidcSecret
-
-	if err := s.store.UpdateSettings(r.Context(), set); err != nil {
-		writeError(w, err)
-		return
-	}
-	s.emit(r.Context(), Event{
-		Type: "settings.changed", Severity: "info", Source: "frontdesk",
-		Message: "Settings updated",
-	})
-	// Re-mask before echoing the saved settings back to the client.
-	if set.AlertAppriseTargets != "" {
-		set.AlertAppriseTargets = alertMaskValue
-	}
-	if set.OidcClientSecret != "" {
-		set.OidcClientSecret = alertMaskValue
-	}
-	writeJSON(w, http.StatusOK, set)
-}
-
-// resolveSecret maps a submitted masked-secret field to the value to store: the
-// mask sentinel preserves the existing stored ciphertext (read back via current),
-// a blank clears it, and any other value is encrypted at rest with the Front Desk
-// master key. Shared by every settings secret (Apprise target, OIDC client secret).
-func (s *Server) resolveSecret(ctx context.Context, submitted string, current func(Settings) string) (string, error) {
-	switch submitted {
-	case alertMaskValue:
-		cur, err := s.store.GetSettings(ctx)
-		if err != nil {
-			return "", err
-		}
-		return current(cur), nil
-	case "":
-		return "", nil
-	default:
-		return auth.EncryptString(submitted, s.masterKey)
-	}
-}
-
-// resolveAlertTarget resolves the Apprise target secret (see resolveSecret).
-func (s *Server) resolveAlertTarget(ctx context.Context, submitted string) (string, error) {
-	return s.resolveSecret(ctx, submitted, func(cur Settings) string { return cur.AlertAppriseTargets })
-}
-
-// getAutoSync returns the automatic config-propagation setup (enabled + the
-// designated primary). The internal drift hash is never included.
-func (s *Server) getAutoSync(w http.ResponseWriter, r *http.Request) {
-	cfg, err := s.store.GetAutoSync(r.Context())
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	writeJSON(w, http.StatusOK, cfg)
-}
-
-// putAutoSync sets the auto-sync toggle and designated primary. Enabling without
-// a primary, or naming a primary that is unknown or has no stored admin token, is
-// rejected: the loop could not authenticate to pull its config, so the choice
-// would silently do nothing.
-//
-// Repointing or clearing an already-configured primary is high-impact (it changes
-// which instance's config gets pushed across the whole fleet), so it is gated on a
-// fresh admin-token confirmation. The bearer may be a passkey/TOTP session token
-// rather than the raw FRONTDESK_TOKEN, so the check happens here against AdminMgr
-// rather than client-side. The first primary selection (none configured yet) and
-// changes that leave the primary untouched (e.g. just toggling enabled) need no
-// confirmation.
-// repointTargetsCurrentPrimary reports whether repointing the fleet primary to
-// candidateID would land on the same physical host that is already the primary,
-// reached under a different URL. It asks the candidate host's own HA self-report
-// (see memberReportsPrimary), so the answer is independent of the member id or
-// URL string. Returns false on the first designation (no current primary), on a
-// same-member-row re-select (a no-op), and it fails open (false) when the
-// candidate cannot be probed, since the admin-token gate still protects the
-// repoint and blocking a legitimate change on a transient read is worse.
-func (s *Server) repointTargetsCurrentPrimary(ctx context.Context, candidateID string) (bool, error) {
-	cur, err := s.store.GetAutoSync(ctx)
-	if err != nil {
-		return false, err
-	}
-	if cur.PrimaryID == "" || cur.PrimaryID == candidateID {
-		return false, nil
-	}
-	m, err := s.store.GetMember(ctx, candidateID)
-	if err != nil {
-		return false, err
-	}
-	token, ok, err := s.store.MemberToken(ctx, candidateID)
-	if err != nil {
-		return false, err
-	}
-	if !ok {
-		return false, nil
-	}
-	isPrimary, _, determined := s.memberIdentity(ctx, m.URL, token)
-	return determined && isPrimary, nil
-}
-
-// instanceAlreadyMember reports whether instanceID belongs to a member other
-// than excludeID. It compares against each member's stored instance_id; for a
-// member whose identity is not yet known (empty stored id, e.g. added before
-// instance identity existed) it probes /api/system once and backfills the
-// learned id, so the check is correct without a separate migration pass. A
-// member that cannot be probed is skipped (it simply cannot be deduped yet); a
-// store read failure is surfaced so the caller can refuse rather than guess.
-//
-// Two simultaneous adds of the same physical instance under different URLs can
-// both create a row before either records its instance_id, so each dedup pass
-// sees the other and both roll back. That resolves toward the safe outcome
-// (neither added, no duplicate persisted); the operator simply retries once.
-func (s *Server) instanceAlreadyMember(ctx context.Context, excludeID, instanceID string) (bool, error) {
-	members, err := s.store.ListMembers(ctx)
-	if err != nil {
-		return false, err
-	}
-	for _, m := range members {
-		if m.ID == excludeID {
-			continue
-		}
-		known := m.InstanceID
-		if known == "" && m.HasToken {
-			token, ok, terr := s.store.MemberToken(ctx, m.ID)
-			if terr == nil && ok {
-				if _, id, identOK := s.memberIdentity(ctx, m.URL, token); identOK && id != "" {
-					known = id
-					if serr := s.store.SetMemberInstanceID(ctx, m.ID, id); serr != nil {
-						debuglog.Warn("frontdesk: could not backfill member instance id", "member", m.ID, "error", serr)
-					}
-				}
-			}
-		}
-		if known == instanceID {
-			return true, nil
-		}
-	}
-	return false, nil
-}
-
-func (s *Server) putAutoSync(w http.ResponseWriter, r *http.Request) {
-	var req struct {
-		Enabled      bool   `json:"enabled"`
-		PrimaryID    string `json:"primary_id"`
-		ConfirmToken string `json:"confirm_token"`
-	}
-	if !decodeJSON(w, r, &req) {
-		return
-	}
-	if req.PrimaryID != "" {
-		m, err := s.store.GetMember(r.Context(), req.PrimaryID)
-		if err != nil {
-			if errors.Is(err, ErrNotFound) {
-				http.Error(w, "that primary is not a known member", http.StatusBadRequest)
-				return
-			}
-			writeError(w, err)
-			return
-		}
-		if !m.HasToken {
-			http.Error(w, "store an admin token for that primary first; auto-sync needs it to read the primary's config", http.StatusBadRequest)
-			return
-		}
-	} else if req.Enabled {
-		http.Error(w, "choose a primary before enabling auto-sync", http.StatusBadRequest)
-		return
-	}
-	// Repointing or clearing an already-configured primary is high-impact (it
-	// changes which instance's config the whole fleet copies), so it requires a
-	// valid admin token. The bearer here may be a passkey/TOTP session token
-	// rather than the raw FRONTDESK_TOKEN, so the operator re-supplies the token
-	// in the body and we check it against AdminMgr. The guard is enforced inside
-	// the same UPDATE that writes (SetAutoSyncGuarded), so there is no
-	// read-modify-write window for a concurrent repoint to slip through.
-	tokenValid := s.adminMgr.Validate(strings.TrimSpace(req.ConfirmToken))
-	// Same-host guard: a valid token authorises changing the primary, but if the
-	// newly-selected host is the same physical instance as the current primary
-	// reached under a different URL (public DNS vs a LAN address), the "change" is
-	// a no-op replacement of the source of truth with itself. Ask the candidate
-	// host its own HA self-report (id/URL-independent) and refuse if it already is
-	// the primary. Only checked on an authorised repoint, so no/wrong-token
-	// attempts short-circuit at the token gate below without probing a member.
-	if tokenValid && req.PrimaryID != "" {
-		same, serr := s.repointTargetsCurrentPrimary(r.Context(), req.PrimaryID)
-		if serr != nil {
-			writeError(w, serr)
-			return
-		}
-		if same {
-			http.Error(w, "you cannot replace the primary with the same host: the selected host is already the fleet primary (reached under a different address)", http.StatusConflict)
-			return
-		}
-	}
-	applied, err := s.store.SetAutoSyncGuarded(r.Context(), req.Enabled, req.PrimaryID, tokenValid)
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	if !applied {
-		// The change would repoint/clear a configured primary without a valid
-		// token (or lost a concurrent repoint). No secrets are logged, only that a
-		// confirmation-required change was refused.
-		debuglog.Warn("frontdesk: refused unconfirmed auto-sync primary change", "to", req.PrimaryID)
-		http.Error(w, "confirm the admin token to change or clear the configured primary", http.StatusForbidden)
-		return
-	}
-	// The guarded write bumped the generation: cancel any pass still importing the
-	// old primary's config before the kick below starts a fresh pass.
-	s.signalRearm()
-	s.emit(r.Context(), Event{
-		Type: "settings.changed", Severity: "info", Source: "frontdesk",
-		Message: fmt.Sprintf("Auto-sync %s", enabledWord(req.Enabled)),
-	})
-	cfg, err := s.store.GetAutoSync(r.Context())
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	// When auto-sync is left on, converge the fleet right away instead of waiting
-	// up to two ticks for the loop: the operator opted in deliberately, so this is
-	// the "sync now" they expect from the toggle. Detached from the request
-	// context (which ends when we respond) but time-bounded so a stuck pass cannot
-	// leak the goroutine. A no-op when nothing has drifted; the loop still owns the
-	// steady-state watch. Disabling (or no primary) never kicks.
-	if cfg.Enabled && cfg.PrimaryID != "" {
-		kickCtx, cancel := context.WithTimeout(context.WithoutCancel(r.Context()), autoSyncKickTimeout)
-		s.bgWG.Add(1)
-		go func() {
-			defer s.bgWG.Done()
-			defer cancel()
-			s.forceAutoSyncNow(kickCtx)
-		}()
-	}
-	writeJSON(w, http.StatusOK, cfg)
-}
-
-func enabledWord(b bool) string {
-	if b {
-		return "enabled"
-	}
-	return "disabled"
-}
-
-// rearmAutoSync clears the last-applied config hash so the auto-sync loop runs a
-// full convergence pass on its next tick. The loop's fast path skips work when the
-// primary's hash is unchanged, so a member that becomes newly syncable (just
-// added, or just given an admin token) would otherwise stay stale until the
-// primary's config next changed. Clearing the marker re-arms the loop; members
-// already matching the primary are still skipped by their own dry-run diff, so the
-// re-run only syncs the one(s) that actually need it. It is a no-op in effect when
-// auto-sync is disabled (the loop reads the marker but does nothing).
-func (s *Server) rearmAutoSync(ctx context.Context) {
-	if err := s.store.RearmAutoSync(ctx); err != nil {
-		debuglog.Warn("frontdesk: re-arm auto-sync after membership change", "error", err)
-		return
-	}
-	s.signalRearm()
-}
-
-// signalRearm wakes every in-flight convergence pass so it cancels synchronously,
-// the instant a rearm/repoint has bumped the auto-sync generation, rather than on
-// the next poll. Call it only after the generation-bumping write has committed.
-// Closing the channel broadcasts to all current waiters; a fresh channel replaces
-// it for the next pass. watchRearm still confirms the generation actually moved
-// before cancelling, so a spurious wake never aborts a valid pass.
-func (s *Server) signalRearm() {
-	s.rearmMu.Lock()
-	close(s.rearmCh)
-	s.rearmCh = make(chan struct{})
-	s.rearmMu.Unlock()
-}
-
-// rearmChan returns the current rearm-broadcast channel for a pass to wait on. It
-// must be captured before the pass reads the generation it guards, so a rearm
-// landing in between still wakes the waiter (the channel is closed, not missed).
-func (s *Server) rearmChan() <-chan struct{} {
-	s.rearmMu.Lock()
-	defer s.rearmMu.Unlock()
-	return s.rearmCh
-}
-
-// ---------------------------------------------------------------------------
-// Events
-// ---------------------------------------------------------------------------
-
-func (s *Server) listEvents(w http.ResponseWriter, r *http.Request) {
-	q := r.URL.Query()
-	f := EventFilter{
-		MemberID: q.Get("member_id"),
-		Type:     q.Get("type"),
-		Severity: q.Get("severity"),
-		Since:    parseRFC3339(q.Get("since")),
-		Until:    parseRFC3339(q.Get("until")),
-		Limit:    clampEventsLimit(atoiDefault(q.Get("limit"), defaultEventsLimit)),
-		Offset:   max(atoiDefault(q.Get("offset"), 0), 0),
-	}
-	evs, total, err := s.store.ListEvents(r.Context(), f)
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	writeJSON(w, http.StatusOK, map[string]any{"events": evs, "total": total})
-}
-
-// ---------------------------------------------------------------------------
-// Status + SSE
-// ---------------------------------------------------------------------------
-
-func (s *Server) traefikStatus(w http.ResponseWriter, _ *http.Request) {
-	writeJSON(w, http.StatusOK, s.poller.Snapshot())
-}
-
-// buildCommit is the source commit SHA this Front Desk binary was built from,
-// stamped at build time via -ldflags -X (see the Makefile / Dockerfile.frontdesk)
-// and surfaced read-only as app_commit so the UI footer can show which commit a
-// `dev` build corresponds to. Defaults to "unknown" for un-stamped builds.
-var buildCommit = "unknown"
-
-// getVersion returns the running build's version and source commit so the UI
-// footer can show which Front Desk build is deployed (and link a `dev` build to
-// its commit on GitHub). app_version is "dev" for un-stamped builds; app_commit
-// is normalized to a short prefix so it reads the same across build paths.
-func (s *Server) getVersion(w http.ResponseWriter, _ *http.Request) {
-	writeJSON(w, http.StatusOK, map[string]string{
-		"app_version": s.version,
-		"app_commit":  util.ShortCommit(buildCommit),
-	})
-}
-
-// getObservability reports which log-export integrations are active, derived
-// read-only from the process environment (LOG_FORMAT, OTEL_EXPORTER_OTLP_*).
-// It mirrors the main server's log_export_* status keys so the Front Desk
-// Observability panel can reflect the same state. Nothing here is runtime-
-// changeable; each integration is enabled by its own environment variable.
-// log_export_metrics reports whether a dedicated scrape token is configured
-// (the /metrics endpoint itself always exists, gated by admin auth otherwise),
-// matching the main server's semantics.
-func (s *Server) getObservability(w http.ResponseWriter, _ *http.Request) {
-	writeJSON(w, http.StatusOK, map[string]bool{
-		"log_export_json":    debuglog.JSONFormat(),
-		"log_export_otel":    otelexport.LogsEnabled(),
-		"log_export_metrics": s.metricsToken != "",
-	})
-}
-
-// sseHeartbeat keeps idle SSE connections alive through proxies.
-const sseHeartbeat = 25 * time.Second
-
-func (s *Server) sse(w http.ResponseWriter, r *http.Request) {
-	flusher, ok := w.(http.Flusher)
-	if !ok {
-		http.Error(w, "streaming unsupported", http.StatusInternalServerError)
-		return
-	}
-	w.Header().Set("Content-Type", "text/event-stream")
-	w.Header().Set("Cache-Control", "no-cache")
-	w.Header().Set("Connection", "keep-alive")
-	w.WriteHeader(http.StatusOK)
-	flusher.Flush()
-
-	ch := s.bus.Subscribe()
-	defer s.bus.Unsubscribe(ch)
-
-	ticker := time.NewTicker(sseHeartbeat)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-r.Context().Done():
-			return
-		case <-ticker.C:
-			if _, err := w.Write([]byte(": keep-alive\n\n")); err != nil {
-				return
-			}
-			flusher.Flush()
-		case ev, ok := <-ch:
-			if !ok {
-				return
-			}
-			data, err := json.Marshal(ev)
-			if err != nil {
-				continue
-			}
-			if _, err := w.Write([]byte("data: " + string(data) + "\n\n")); err != nil {
-				return
-			}
-			flusher.Flush()
-		}
-	}
-}
-
-// ---------------------------------------------------------------------------
-// TOTP-enabled cache (mirrors the main server's cache so the gate stays DB-free
-// on the hot path)
-// ---------------------------------------------------------------------------
-
-// totpStatusReader is the one method of *totp.Repository the cache depends on.
-// It is an interface so the fail-closed behaviour on a read error is testable
-// without a live database.
-type totpStatusReader interface {
-	IsEnabled(ctx context.Context) (bool, error)
-}
-
-type totpEnabledCache struct {
-	repo totpStatusReader
-	val  atomic.Bool
-}
-
-func newTotpEnabledCache(repo totpStatusReader) *totpEnabledCache {
-	c := &totpEnabledCache{repo: repo}
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-	enabled, err := repo.IsEnabled(ctx)
-	if err != nil {
-		// Fail closed: treat as enabled so a startup DB blip cannot silently
-		// weaken the gate.
-		debuglog.Error("frontdesk: seeding TOTP-enabled cache failed, failing closed", "error", err)
-		enabled = true
-	}
-	c.val.Store(enabled)
-	return c
-}
-
-func (c *totpEnabledCache) Enabled() bool { return c.val.Load() }
-
-func (c *totpEnabledCache) Refresh(ctx context.Context) {
-	enabled, err := c.repo.IsEnabled(ctx)
-	if err != nil {
-		// Fail closed, matching the main server's RefreshTotpEnabled: a failed
-		// re-read must never leave a stale "disabled" cached, which would let a
-		// raw FRONTDESK_TOKEN through as a full session after TOTP was enabled.
-		debuglog.Error("frontdesk: refreshing TOTP-enabled cache failed, failing closed", "error", err)
-		c.val.Store(true)
-		return
-	}
-	c.val.Store(enabled)
-}
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-// emit persists a control-plane event and publishes it on the SSE bus.
-func (s *Server) emit(ctx context.Context, e Event) {
-	stored, err := s.store.InsertEvent(ctx, e)
-	if err != nil {
-		debuglog.Warn("frontdesk: persist event", "type", e.Type, "error", err)
-		stored = e
-	}
-	s.bus.Publish(busEvent(stored))
-}
-
-// busEvent maps a stored Front Desk Event to a bus event. When the event concerns
-// a member, its MemberID is copied into the metadata as "member_id" (on a copy, so
-// the persisted metadata is untouched) so the alert dispatcher debounces per
-// member. Shared by Server.emit and Poller.recordEvent.
-func busEvent(e Event) events.Event {
-	meta := e.Metadata
-	if e.MemberID != "" {
-		meta = make(map[string]any, len(e.Metadata)+1)
-		for k, v := range e.Metadata {
-			meta[k] = v
-		}
-		meta["member_id"] = e.MemberID
-	}
-	return events.Event{
-		ID: e.ID, Type: e.Type, Severity: e.Severity, Source: e.Source,
-		Message: e.Message, Metadata: meta, Timestamp: e.CreatedAt,
-	}
-}
-
-// RunAlerts runs the outbound Apprise dispatcher until ctx is cancelled. Started
-// as a goroutine in cmd/frontdesk; best-effort, never blocks request serving.
-func (s *Server) RunAlerts(ctx context.Context) { s.alertDisp.Run(ctx) }
-
-// alertEvents serves the Front Desk alert catalog so the UI renders its picker.
-func (s *Server) alertEvents(w http.ResponseWriter, _ *http.Request) {
-	writeJSON(w, http.StatusOK, fdCatalog)
-}
-
-// alertStatus reports whether the configured apprise-api is reachable. A
-// reachable host is not enough: if the stored target cannot be decrypted (master
-// key rotated, ciphertext corrupted) every dispatch fails silently, so that is
-// surfaced as unhealthy with a reason rather than a falsely green pill.
-func (s *Server) alertStatus(w http.ResponseWriter, r *http.Request) {
-	st, err := s.alertDisp.Probe(r.Context())
-	if err != nil {
-		writeError(w, err)
-		return
-	}
-	if st.Configured {
-		if set, gerr := s.store.GetSettings(r.Context()); gerr == nil {
-			switch set.AlertAppriseTargets {
-			case "":
-				// A reachable apprise-api with no target still cannot deliver, so it
-				// must not show a green pill.
-				st.Healthy = false
-				st.Detail = "no notification target configured"
-			default:
-				if _, derr := auth.DecryptString(set.AlertAppriseTargets, s.masterKey); derr != nil {
-					st.Healthy = false
-					st.Detail = "stored target cannot be decrypted (master key rotated?)"
-				}
-			}
-		}
-	}
-	writeJSON(w, http.StatusOK, st)
-}
-
-// alertTest sends a test notification to the configured target(s). A delivery or
-// configuration failure is reported as 502 with the reason, so the UI can show it.
-func (s *Server) alertTest(w http.ResponseWriter, r *http.Request) {
-	if err := s.alertDisp.TestSend(r.Context()); err != nil {
-		writeJSON(w, http.StatusBadGateway, map[string]string{"error": err.Error()})
-		return
-	}
-	w.WriteHeader(http.StatusNoContent)
-}
-
-func writeJSON(w http.ResponseWriter, status int, v any) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(status)
-	if err := json.NewEncoder(w).Encode(v); err != nil {
-		debuglog.Error("frontdesk: encode response", "error", err)
-	}
-}
-
-// writeError maps store errors to HTTP status codes.
-// writeCodedError writes a JSON error body carrying a stable machine-readable
-// code alongside the human message, so the frontend can route on the code rather
-// than matching translatable English text. Plain-text writeError is kept for the
-// many endpoints that need no client-side discrimination.
-func writeCodedError(w http.ResponseWriter, status int, code, msg string) {
-	writeJSON(w, status, map[string]string{"code": code, "error": msg})
-}
-
-func writeError(w http.ResponseWriter, err error) {
-	switch {
-	case errors.Is(err, ErrNotFound):
-		http.Error(w, err.Error(), http.StatusNotFound)
-	case errors.Is(err, ErrValidation), errors.Is(err, ErrDuplicateURL), errors.Is(err, ErrInsecureURL):
-		http.Error(w, err.Error(), http.StatusBadRequest)
-	default:
-		debuglog.Error("frontdesk: request failed", "error", err)
-		http.Error(w, "internal error", http.StatusInternalServerError)
-	}
-}
-
-// decodeJSON decodes the request body, writing a 400 and returning false on
-// failure.
-func decodeJSON(w http.ResponseWriter, r *http.Request, v any) bool {
-	if err := json.NewDecoder(r.Body).Decode(v); err != nil {
-		http.Error(w, "invalid JSON body", http.StatusBadRequest)
-		return false
-	}
-	return true
-}
-
-func atoiDefault(s string, def int) int {
-	if s == "" {
-		return def
-	}
-	n, err := strconv.Atoi(s)
-	if err != nil {
-		return def
-	}
-	return n
-}
-
-// Event listing page-size bounds. A request with no/blank limit gets the
-// default; a non-positive limit would otherwise disable the store's LIMIT clause
-// (unbounded query), and an over-large one could return the whole table, so both
-// ends are clamped here.
-const (
-	defaultEventsLimit = 100
-	maxEventsLimit     = 500
-)
-
-// clampEventsLimit forces an events page size into [1, maxEventsLimit].
-func clampEventsLimit(n int) int {
-	if n < 1 {
-		return defaultEventsLimit
-	}
-	if n > maxEventsLimit {
-		return maxEventsLimit
-	}
-	return n
-}
-
-// parseRFC3339 parses an RFC3339 timestamp from a query value, returning the
-// zero time (which EventFilter treats as "no bound") when empty or malformed.
-func parseRFC3339(s string) time.Time {
-	if s == "" {
-		return time.Time{}
-	}
-	t, err := time.Parse(time.RFC3339, s)
-	if err != nil {
-		return time.Time{}
-	}
-	return t
-}
-
-// spaHandler serves the embedded single-page app, falling back to index.html for
-// client-side routes (any path without a file extension that is not found).
-func spaHandler(ui fs.FS) http.Handler {
-	fileServer := http.FileServer(http.FS(ui))
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// fs.ValidPath + the embedded FS are the traversal boundary: "../" or an
-		// absolute name is rejected here and falls through to the SPA index, and
-		// http.FileServer additionally cleans the path it serves. Only serve a
-		// concrete asset when it exists and the name is valid.
-		name := strings.TrimPrefix(r.URL.Path, "/")
-		if name != "" && fs.ValidPath(name) {
-			if _, err := fs.Stat(ui, name); err == nil {
-				fileServer.ServeHTTP(w, r)
-				return
-			}
-		}
-		// Root, invalid, or unknown path: serve index.html so the SPA router can
-		// handle the route client-side.
-		r2 := r.Clone(r.Context())
-		r2.URL.Path = "/"
-		fileServer.ServeHTTP(w, r2)
-	})
 }
